Fixes 29634: Consolidate tooltip helpers and test
